moved some files into root level dir, further refactored app.py
