+++ scripts/run_indexing.py
